// client/src/components/SubmissionStats/types.ts
export type StatsData = {
  mode_300?: {
    count: number;
    averageScore: number;
    maxScore: number;
    recentDate: string;
    averageDuration: number;
  };
  mode_1000?: {
    count: number;
    averageScore: number;
    maxScore: number;
    recentDate: string;
    averageDuration: number;
    averageSessionCount: number;
  };
};